getStateFromIdPath
