+++ frontend/src/metabase/common/components/Pickers/utils.ts
@@ -1,7 +1,17 @@
 import { PERSONAL_COLLECTIONS } from "metabase/entities/collections/constants";
-import type { CollectionId } from "metabase-types/api";
+import { isNullOrUndefined } from "metabase/lib/types";
+import type {
+  CollectionId,
+  CollectionItemModel,
+  ListCollectionItemsRequest,
+} from "metabase-types/api";
 
-import type { CollectionPickerItem } from "./CollectionPicker";
+import type { PickerState } from "../EntityPicker";
+
+import type {
+  CollectionPickerItem,
+  CollectionPickerStatePath,
+} from "./CollectionPicker";
 
 export const getCollectionIdPath = (
   collection: Pick<
@@ -40,5 +50,79 @@ export const getCollectionIdPath = (
     return ["personal", ...pathFromRoot, id];
   } else {
     return ["root", ...pathFromRoot, id];
+  }
+};
+
+export const getStateFromIdPath = ({
+  idPath,
+  namespace,
+  models,
+}: {
+  idPath: CollectionId[];
+  namespace?: "snippets";
+  models: CollectionItemModel[];
+}): CollectionPickerStatePath => {
+  const statePath: PickerState<
+    CollectionPickerItem,
+    ListCollectionItemsRequest
+  > = [
+    {
+      selectedItem: {
+        name: "",
+        model: "collection",
+        id: idPath[0],
+        here: ["collection"],
+        below: ["collection"],
+      },
+    },
+  ];
+
+  idPath.forEach((id, index) => {
+    const { entityId: nextLevelId, model: nextLevelModel } = resolveEntityId(
+      idPath[index + 1],
+    );
+
+    const { entityId, model: entityModel } = resolveEntityId(id);
+
+    statePath.push({
+      query: {
+        id: entityId,
+        models: ["collection", ...models],
+        namespace,
+      },
+      entity: entityModel,
+      selectedItem: nextLevelId
+        ? {
+            name: "",
+            model: nextLevelModel,
+            id: nextLevelId,
+            here: ["collection"],
+            below: ["collection"],
+          }
+        : null,
+    });
+  });
+
+  return statePath;
+};
+
+const resolveEntityId = (
+  id: CollectionId,
+): {
+  model: "collection" | "dashboard";
+  entityId: CollectionId;
+} => {
+  if (typeof id === "string" || isNullOrUndefined(id)) {
+    return {
+      entityId: id,
+      model: "collection",
+    };
+  } else {
+    const isDashboard = id < 0;
+
+    return {
+      entityId: Math.abs(id),
+      model: isDashboard ? "dashboard" : "collection",
+    };
   }
 };
